CB-5332 fix multiple termination trigger
