+++ core/src/main/java/com/sequenceiq/cloudbreak/job/StackStatusCheckerJob.java
@@ -46,6 +46,7 @@ import com.sequenceiq.cloudbreak.domain.stack.instance.InstanceMetaData;
 import com.sequenceiq.cloudbreak.logger.MDCBuilder;
 import com.sequenceiq.cloudbreak.service.cluster.ClusterApiConnectors;
 import com.sequenceiq.cloudbreak.service.cluster.ClusterService;
+import com.sequenceiq.cloudbreak.service.cluster.flow.ClusterFlowService;
 import com.sequenceiq.cloudbreak.service.environment.credential.CredentialConverter;
 import com.sequenceiq.cloudbreak.service.stack.InstanceMetaDataService;
 import com.sequenceiq.cloudbreak.service.stack.StackService;
@@ -69,6 +70,9 @@ public class StackStatusCheckerJob extends StatusCheckerJob {
 
     @Inject
     private ClusterService clusterService;
+
+    @Inject
+    private ClusterFlowService clusterFlowService;
 
     @Inject
     private ClusterApiConnectors clusterApiConnectors;
@@ -142,7 +146,7 @@ public class StackStatusCheckerJob extends StatusCheckerJob {
         Set<String> failedNodeNames = failedInstances.stream()
                 .map(InstanceMetaData::getDiscoveryFQDN)
                 .collect(toSet());
-        clusterService.reportHealthChange(stack.getResourceCrn(), failedNodeNames, newHealtyHostNames);
+        clusterFlowService.reportHealthChange(stack.getResourceCrn(), failedNodeNames, newHealtyHostNames);
         if (failedNodeNames.size() > 0) {
             clusterService.updateClusterStatusByStackId(stack.getId(), Status.AMBIGUOUS);
         } else if (EnumSet.of(Status.AMBIGUOUS, Status.STOPPED).contains(stack.getCluster().getStatus())) {
