+++ core/src/test/java/com/sequenceiq/cloudbreak/job/StackStatusCheckerJobTest.java
@@ -35,6 +35,7 @@ import com.sequenceiq.cloudbreak.domain.stack.cluster.Cluster;
 import com.sequenceiq.cloudbreak.domain.stack.instance.InstanceMetaData;
 import com.sequenceiq.cloudbreak.service.cluster.ClusterApiConnectors;
 import com.sequenceiq.cloudbreak.service.cluster.ClusterService;
+import com.sequenceiq.cloudbreak.service.cluster.flow.ClusterFlowService;
 import com.sequenceiq.cloudbreak.service.environment.credential.CredentialConverter;
 import com.sequenceiq.cloudbreak.service.stack.InstanceMetaDataService;
 import com.sequenceiq.cloudbreak.service.stack.StackService;
@@ -81,6 +82,9 @@ public class StackStatusCheckerJobTest {
 
     @Mock
     private ClusterService clusterService;
+
+    @Mock
+    private ClusterFlowService clusterFlowService;
 
     @Mock
     private ClusterApi clusterApi;
@@ -200,7 +204,7 @@ public class StackStatusCheckerJobTest {
         setupForCM();
         underTest.executeInternal(jobExecutionContext);
 
-        verify(clusterService, times(0)).reportHealthChange(anyString(), anySet(), anySet());
+        verify(clusterFlowService, times(0)).reportHealthChange(anyString(), anySet(), anySet());
         verify(instanceStateQuery, times(1)).getCloudVmInstanceStatuses(any(), any(), any());
     }
 
@@ -212,7 +216,7 @@ public class StackStatusCheckerJobTest {
         when(clusterApi.clusterStatusService()).thenReturn(clusterStatusService);
         underTest.executeInternal(jobExecutionContext);
 
-        verify(clusterService, times(1)).reportHealthChange(any(), anySet(), anySet());
+        verify(clusterFlowService, times(1)).reportHealthChange(any(), anySet(), anySet());
         verify(instanceStateQuery, times(1)).getCloudVmInstanceStatuses(any(), any(), any());
     }
 }
